+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] / 2025-07-30
+### Features
+- Support install progress bar.
+### Updates
+- Update `CreateDataColumns` to support more values in the `DataTable`.
+- Add `ConsoleTaskExtension` to wait task to finish and show progress bar.
+- Update `AppBundleUtils` with `FindAppBundle` to find `AppBundle` by app name or bundle folder.
+- Update `CACHE_TOTAL_MINUTES` in `DownloadUtils` to 1 second.
+- Remove `ricaun.Revit.Installation` package reference.
+
 ## [1.0.1] / 2025-05-31
 ### Features
 - Support `Autodesk` native bundle with admin privileges.
@@ -31,5 +41,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `DownloadUtils` to support local file.
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.0.2]: ../../compare/1.0.1...1.0.2
 [1.0.1]: ../../compare/1.0.0...1.0.1
 [1.0.0]: ../../compare/1.0.0